Updated bulk routes. Changed UX for bulk well aquifer correlation Vue
