+++ app/frontend/src/router.js
@@ -8,7 +8,8 @@ import { store } from './store/index.js'
 import AquiferSearch from '@/aquifers/components/Search'
 import AquiferView from '@/aquifers/components/View'
 import AquiferNew from '@/aquifers/components/New'
-import BulkAquiferUpload from '@/aquifers/components/BulkUpload'
+import BulkHome from '@/bulk/components/BulkHome.vue'
+import BulkWellAquiferCorrelation from '@/bulk/components/BulkWellAquiferCorrelation.vue'
 
 import WellSearch from '@/wells/views/WellSearch.vue'
 import WellDetail from '@/wells/views/WellDetail.vue'
@@ -63,9 +64,14 @@ const router = new Router({
       }
     },
     {
-      path: '/aquifers/bulk',
-      component: BulkAquiferUpload,
-      name: 'BulkAquiferUpload',
+      path: '/bulk/',
+      component: BulkHome,
+      name: 'bulk-home'
+    },
+    {
+      path: '/bulk/well-aquifer-correlation',
+      component: BulkWellAquiferCorrelation,
+      name: 'bulk-well-aquifer-correlation',
       beforeEnter: AuthGuard,
       meta: {
         edit: true,
